Introduce option not to display definitions in repl driver
The reason for this change is that currently, statements in the
JSR223 script engine are evaluated twice. Once when the "run" method
is called, and once when the value of the result is extracted.
The first one needs not be, it exists solely due to displayDefinitions
being called. If we put it behind a flag, the problem is solved.